feat(livechart): add custom styling object for livechart view paints
New class LiveChartStyle.kt used to style the LiveChartView through the public api setLiveChartStyle(style: LiveChartStyle).

It removes as default the baseline conditional coloring and falls back to the main color in the LiveChartStyle. Baseline conditional coloring can be toggled with the public api: drawBaselineConditionalColor() , and positive/negative colors set in the style object.

Closes #7