feat(overlay): overlay and view combined into one component
The touch overlay and view are now exposed through the LiveChart frame layout. Public Apis are exposed through this class. Overlay and bound label bugs with second datasets are also fixed.

The LiveChart class exposes two new APIs. 1. A callback to get the current data point in the touch gesture 2. disableTouchOverlay() to hide and stop the touch overlay from binding to the dataset.